requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
azure-cognitiveservices-speech==1.26.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.92.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
multidict==6.0.4
numpy==1.24.2
openai==0.27.0
pycparser==2.21
pydantic==1.10.5
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
sounddevice==0.4.6
starlette==0.25.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
# uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
